progressao/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
crispy-bootstrap5==2025.4
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.1
django-braces==1.17.0
django-cleanup==9.0.0
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.58.0
html5lib==1.1
idna==3.10
lxml==5.4.0
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
